requirements.txt
'from
.
BeautifulSoup
argparse
cProfile
cgi
codecs
copy
datetime
feedparser
from
glob
html.parser
import
json
math
multiprocessing
nested
networkx
numpy
operator
os
random
re
rfc822
scipy
scipy.cluster.vq
sgmllib
stemming.porter2
sys
telnetlib
theseus.processor.theseus
time
timeit
tkinter
types
urllib.error
urllib.parse
urllib.request
zlib
